+++ mne/datasets/bnci/__init__.py
@@ -1,0 +1,3 @@
+"""BNCI Datasets."""
+
+from .bnci import load_data